Backwards compatibility for adsorbate db
https://github.com/Open-Catalyst-Project/Open-Catalyst-Dataset/pull/67
added support for reaction strings in adsorbate database entries. Tuple
unpacking in the Adsorbate initializer assumes length 4 (after the
reaction string was added). Older databases will be missing the reaction
string and the initializer will raise when they are used. This makes the
initializer backwards compatible with older databases, saving the
reaction string only if present.